Move validator module to client-modules folder
